General bug fixes and improvements.
Added a delay to the start of the demo program to ensure it boots OK.
(From experiments, the display seems to need up to 4 seconds to come
fully online.)
Prevented the library class from resetting the display on construction.
I've found that the hard reset is somewhat unreliable. It seems better
to have reset available optionally via a function.
Added a flush() function so that the programmer can wait until previous
commands have finished. This is potentially very useful when sending
lots of commands which take time, such as clearing the screen or filling
large shapes.
Changed formatting of library files to use the more common 4-space
indents instead of the Arduino standard of 2, mainly because of personal
preference. The example program still uses 2 though since people are
likely to edit it in the Arduino IDE.